+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- 6 new font
+	1. small_fancy
+	2. boom
+	3. cross_jesus
+	4. dark_tattoo
+	5. lucifer
+	6. ascii_roman
 ### Changed
 - Test system modified
 - `README.md` modified